refactor: petition service interface 삭제
